+++ packages/mui-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/mui-components
 
+## 3.0.28
+
+### Patch Changes
+
+- Update styles
+
 ## 3.0.27
 
 ### Patch Changes